# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Amrutam Telemedicine API
  version: 1.0.0
  description: |
    Production-grade REST API for an Ayurveda telemedicine platform.

    ## Authentication
    Protected endpoints require a JWT bearer token:
    ```
    Authorization: Bearer <access_token>
    ```
    Obtain tokens via `POST /api/v1/auth/login`.

    ## Idempotency
    `POST /api/v1/appointments` requires an `Idempotency-Key` header (UUID recommended).

    ## Response Envelope
    Success: `{ success, data, meta, requestId, timestamp }`
    Error: `{ success: false, code, message, details, requestId, timestamp }`
  contact:
    name: Amrutam Engineering
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from POST /auth/login
    idempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
      description: Client-generated UUID for idempotent booking
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data, requestId, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data: {}
        meta:
          type: object
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    ErrorEnvelope:
      type: object
      required: [success, code, message, requestId, timestamp]
      properties:
        success:
          type: boolean
          const: false
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          nullable: true
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
          example: patient@example.com
        password:
          type: string
          minLength: 8
          example: Password123!
        firstName:
          type: string
          example: Rahul
        lastName:
          type: string
          example: Sharma
        phone:
          type: string
          example: "919876543210"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: patient@amrutam.test
        password:
          type: string
          example: Password123!
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        timezone:
          type: string
          example: Asia/Kolkata
        language:
          type: string
          example: en
...
  responses:
    Unauthorized:
      description: Invalid or missing bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: UNAUTHORIZED
            message: Invalid or missing bearer token
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    Forbidden:
      description: Insufficient role or resource access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: FORBIDDEN
            message: Insufficient role or resource access
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: VALIDATION_ERROR
            message: Request validation failed
            details: ['doctorId must be a UUID']
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: FORBIDDEN
            message: Rate limit exceeded
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: INTERNAL_ERROR
            message: An unexpected error occurred
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Request correlation ID for distributed tracing
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new patient account
      description: Creates a patient user with bcrypt-hashed password, profile, and issues initial JWT token pair. Audited as REGISTER.
      security: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: patient@example.com
              password: Password123!
              firstName: Rahul
              lastName: Sharma
              phone: "919876543210"
      responses:
        '201':
          description: Account created with tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  user:
                    id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    email: patient@example.com
                    roles: [Patient]
                  accessToken: eyJhbGciOiJIUzI1NiIs...
                  refreshToken: eyJhbGciOiJIUzI1NiIs...
                  expiresIn: 15m
                meta: {}
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                success: false
                code: CONFLICT
                message: Email already registered
                details: null
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: Validates credentials, updates lastLoginAt, audits LOGIN or FAILED_LOGIN, returns access + refresh token pair.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: patient@amrutam.test
              password: Password123!
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              example:
                success: true
                data:
                  user:
                    id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    email: patient@amrutam.test
                    roles: [Patient]
                  accessToken: eyJhbGciOiJIUzI1NiIs...
                  refreshToken: eyJhbGciOiJIUzI1NiIs...
                  expiresIn: 15m
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Invalid credentials
          content:
            application/json:
              example:
                success: false
                code: INVALID_CREDENTIALS
                message: Invalid email or password
                details: null
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '403':
          description: Account inactive
          content:
            application/json:
              example:
                success: false
                code: ACCOUNT_INACTIVE
                message: Account is not active
                details: null
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token
      description: Verifies refresh JWT, revokes stored hash, issues new access + refresh pair. Prevents refresh token reuse.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              refreshToken: eyJhbGciOiJIUzI1NiIs...
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              example:
                success: true
                data:
                  accessToken: eyJhbGciOiJIUzI1NiIs...
                  refreshToken: eyJhbGciOiJIUzI1NiIs...
                  expiresIn: 15m
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token and end session
      description: Marks refresh token as revoked in database. Requires valid access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              example:
                success: true
                data:
                  loggedOut: true
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      description: Returns authenticated user profile and roles.
      responses:
        '200':
          description: Profile retrieved
          content:
            application/json:
              example:
                success: true
                data:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  email: patient@amrutam.test
                  roles: [Patient]
                  profile:
                    firstName: Rahul
                    lastName: Sharma
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Auth]
      summary: Update current user profile
      description: Updates profile fields. Audited as PROFILE_UPDATE.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              firstName: Rahul
              timezone: Asia/Kolkata
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              example:
                success: true
                data:
                  firstName: Rahul
                  timezone: Asia/Kolkata
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/dashboard:
    get:
      tags: [Admin]
      summary: Admin dashboard metrics
      description: Aggregated KPIs cached in Redis (60s TTL). Requires Admin or Super Admin role.
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              example:
                success: true
                data:
                  totalAppointments: 1250
                  todayBookings: 42
                  activeDoctors: 18
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/analytics:
    get:
      tags: [Admin]
      summary: Analytics reporting
      description: Time-series appointment and revenue metrics by period.
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly, custom]
            default: daily
        - name: fromDate
          in: query
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              example:
                success: true
                data:
                  period: daily
                  appointments: []
                  revenue: []
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/audit:
    get:
      tags: [Admin]
      summary: Immutable audit history
      description: Filterable audit log for compliance and incident investigation.
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: resourceType
          in: query
          schema:
            type: string
        - name: fromDate
          in: query
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          schema:
            type: string
            format: date-time
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Audit entries
          content:
            application/json:
              example:
                success: true
                data:
                  items:
                    - action: APPOINTMENT_BOOKED
                      userId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/search:
    get:
      tags: [Admin]
      summary: Global admin search
      description: Search patients, doctors, and appointments by keyword.
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
        - name: types
          in: query
          schema:
            type: string
          description: Comma-separated types (patient, doctor, appointment)
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Search results
          content:
            application/json:
              example:
                success: true
                data:
                  results: []
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...